mcp-server/requirements.txt
asyncpg==0.31.0
fastapi==0.134.0
fastmcp==3.0.2
openai==2.24.0
python-dotenv==1.2.1
uvicorn==0.41.0
